refactor: Move search field constants to custom_search_delegate.dart as they are only used in that file
