+++ lib/src/constants/constants.dart
@@ -1,13 +1,1 @@
-import 'package:flutter/material.dart';
-
-const kSearchFieldTextStyle = TextStyle(
-  fontFamily: 'Poppins',
-  fontSize: 18.0,
-);
-
-final kSearchFieldBorder = OutlineInputBorder(
-  borderRadius: BorderRadius.circular(50.0),
-  borderSide: BorderSide.none,
-);
-
 const prayerTimeShedulerTask = 'me.naffah.namaadhu_app.prayerTimeSchedulerTask';
